changed int to long long int in lucas theorem

+++ ncr_mod_p/lucas.cpp
@@ -12,28 +12,28 @@
 #define endl "\n"
 #define mod 997
 using namespace std;
-vector<vector<int>> dp;
+vector<vector<ll>> dp;
 
-void generate_dp (int a)
+void generate_dp (ll a)
 {
-    dp.resize(a+1,vector<int> (a+1));
-    for (int i=1;i<=a;i++)
+    dp.resize(a+1,vector<ll> (a+1));
+    for (ll i=1;i<=a;i++)
     {
         dp[i][0]=1;
         dp[0][i]=0;
     }
     dp[0][0]=1;
 
-    for (int i=1;i<=a;i++)
+    for (ll i=1;i<=a;i++)
     {
-        for (int j=1;j<=a;j++)
+        for (ll j=1;j<=a;j++)
         {
             dp[i][j]=(dp[i-1][j-1]+dp[i-1][j])%mod;
         }
     }
 }
 
-int ncr_lucas (int n, int r)
+ll ncr_lucas (ll n, ll r)
 {
     if ((r==0)||(r==n))
     {
@@ -41,8 +41,8 @@ int ncr_lucas (int n, int r)
     }
     else
     {
-        int a=n%mod;
-        int b=r%mod;
+        ll a=n%mod;
+        ll b=r%mod;
 
         return (ncr_lucas(n/mod,r/mod)*dp[a][b])%mod;
     }
@@ -54,12 +54,12 @@ int main()
     cin.tie(0);
     cout.precision(10);
 
-    int tt;
+    ll tt;
     cin >> tt;
     generate_dp(mod);
-    for (int yy=0;yy<=tt-1;yy++)
+    for (ll yy=0;yy<=tt-1;yy++)
     {
-        int n,r;
+        ll n,r;
         cin >> n >> r;
         cout << ncr_lucas(n,r) << endl;
     }
